requirements.txt
# Core ML/DL Dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Mamba State Space Model
mamba-ssm>=2.0.0

# Experiment Tracking
wandb>=0.15.0

# Configuration Management
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Development Tools (Optional)
jupyterlab>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Testing (Optional)
pytest>=7.4.0
pytest-cov>=4.1.0
